+++ Project/decorators/trimmer_decorator.rb
